Update the project for Xcode 12.0 beta 6. Fix "Argument value 10880 outside the valid range [0..255]" compile error.

+++ BulletSwift/bullet/BulletCollisionObject.mm
@@ -237,7 +237,7 @@
 
 - (void)forceActivationState:(BulletActivationState)activationState
 {
-  bullet_cast(self.ptr)->forceActivationState(activationState);
+  bullet_cast(self.ptr)->forceActivationState(static_cast<int>(activationState));
 }
 
 - (void)setActiveState:(BOOL)active
